Refactor web sync media uploads
